# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AgentFi API
  version: 0.1.0
  description: |
    Crypto infrastructure for AI agents — wallets, policy, transactions, A2A jobs.


    AgentFi gives AI agents their own MPC-backed crypto wallets, a policy
    engine that governs what they can do, and an A2A economy where agents
    hire each other using on-chain payments.

    This document mirrors the human-readable [api-reference.md](../api-reference.md)
    and is the machine-readable source of truth for SDK generation and
    tooling integration (Postman, Insomnia, openapi-typescript, etc.).

    **Authentication**
    - Agent endpoints: `x-api-key: agfi_live_<hex>` (an Agent API key)
    - Operator actions — agent registration and loosening an agent's policy:
      `x-api-key: <API_SECRET>` (the operator's shared secret)
    - Admin endpoints: `x-admin-secret: <ADMIN_SECRET>`

    **Policy authority.** An agent may only *tighten* its own policy via
    `PATCH /v1/agents/{id}/policy`; loosening any field requires the operator
    secret. See that operation for the per-field rules.

    **Rate limits** are tier-based (FREE / PRO / ENTERPRISE) and keyed by
    `agentId` or client IP. See the Billing section.

    **Errors.** Business refusals are `{ error, code?, ...details }` with a
    stable `code` where the operation documents one. An unexpected server
    error is always `{ "error": "Internal error", "code": "INTERNAL_ERROR",
...
    keyed URLs, internal hosts, stack frames or file paths.
...
servers:
...
security:
  - agentApiKey: []
paths:
...
  /v1/agents/me:
    get:
      tags: [Agents]
      summary: Current agent (resolved from API key)
      responses:
        '200':
          description: Agent profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /v1/transactions/simulate:
    post:
      tags: [Transactions]
      summary: Dry-run a Uniswap V3 swap
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SimulateSwapRequest' }
      responses:
        '200':
          description: Simulation outcome.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimulationResult' }
        '422': { $ref: '#/components/responses/SimulationFailed' }
        '503': { $ref: '#/components/responses/SimulationUnavailable' }
  /v1/transactions/swap:
    post:
      tags: [Transactions]
      summary: Execute a Uniswap V3 swap
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExecuteSwapRequest' }
      responses:
        '202': { $ref: '#/components/responses/TransactionAccepted' }
        '403': { $ref: '#/components/responses/PolicyViolation' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '422': { $ref: '#/components/responses/SimulationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/SimulationUnavailable' }
  /v1/transactions/transfer:
    post:
      tags: [Transactions]
      summary: ETH or ERC-20 transfer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransferRequest' }
      responses:
        '202': { $ref: '#/components/responses/TransactionAccepted' }
        '403': { $ref: '#/components/responses/PolicyViolation' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '422': { $ref: '#/components/responses/SimulationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/SimulationUnavailable' }
  /v1/transactions/deposit:
    post:
      tags: [Transactions]
      summary: Supply to Aave V3
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssetAmountRequest' }
      responses:
        '202': { $ref: '#/components/responses/TransactionAccepted' }
        '403': { $ref: '#/components/responses/PolicyViolation' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '422': { $ref: '#/components/responses/SimulationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/SimulationUnavailable' }
  /v1/transactions/withdraw:
    post:
      tags: [Transactions]
      summary: Withdraw from Aave V3
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssetAmountRequest' }
      responses:
        '202': { $ref: '#/components/responses/TransactionAccepted' }
        '403': { $ref: '#/components/responses/PolicyViolation' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/SimulationUnavailable' }
  /v1/transactions/supply-compound:
    post:
      tags: [Transactions]
      summary: Supply to Compound V3 (Comet USDC market)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssetAmountRequest' }
      responses:
        '202': { $ref: '#/components/responses/TransactionAccepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/PolicyViolation' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/SimulationUnavailable' }
  /v1/transactions/withdraw-compound:
    post:
      tags: [Transactions]
      summary: Withdraw from Compound V3
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssetAmountRequest' }
      responses:
        '202': { $ref: '#/components/responses/TransactionAccepted' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/SimulationUnavailable' }
  /v1/transactions/deposit-erc4626:
    post:
      tags: [Transactions]
      summary: Deposit into any ERC-4626 vault
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Erc4626Request' }
      responses:
        '202': { $ref: '#/components/responses/TransactionAccepted' }
        '403': { $ref: '#/components/responses/PolicyViolation' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/SimulationUnavailable' }
  /v1/transactions/withdraw-erc4626:
    post:
      tags: [Transactions]
      summary: Withdraw from any ERC-4626 vault
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Erc4626Request' }
      responses:
        '202': { $ref: '#/components/responses/TransactionAccepted' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/SimulationUnavailable' }
  /v1/transactions/swap-curve:
    post:
      tags: [Transactions]
      summary: Swap on a Curve StableSwap pool
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CurveSwapRequest' }
      responses:
        '202': { $ref: '#/components/responses/TransactionAccepted' }
        '403': { $ref: '#/components/responses/PolicyViolation' }
        '422': { $ref: '#/components/responses/SimulationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/SimulationUnavailable' }
...
  /v1/transactions/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      tags: [Transactions]
      summary: Transaction status (owner only)
      responses:
        '200':
          description: Full transaction record.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Transaction' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /mcp/messages:
    post:
      tags: [MCP]
      summary: JSON-RPC message handler (paired with /mcp/sse)
      parameters:
        - in: query
          name: sessionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202': { description: Accepted. }
components:
  securitySchemes:
    agentApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Agent-scoped API key (`agfi_live_<hex>`).
    operatorApiKey:
      type: apiKey
      in: header
      name: x-api-key
      description: Operator shared secret. Accepted by `POST /v1/agents` and `PATCH /v1/agents/{id}/policy` (may loosen any agent's policy).
    adminSecret:
      type: apiKey
      in: header
      name: x-admin-secret
      description: Operator admin secret. Required for `/admin/*`.
...
  responses:
    BadRequest:
      description: Validation failed or malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Access denied (wrong agent or insufficient permissions).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    IdempotencyConflict:
      description: The `idempotencyKey` is in use by another agent.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    PolicyViolation:
      description: Action blocked by the agent's operational policy.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    SimulationFailed:
      description: On-chain simulation reverted or produced an error before broadcast.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    SimulationUnavailable:
      description: |
        No usable simulation could authorise the transaction, so it was refused.
        Body: `{ "error": "Simulation unavailable in production" }`.

        In `production`/`staging` the backend never acts on a development mock
        simulation: Tenderly is used when configured, otherwise a real `eth_call`
        (`estimateGas`) dry-run. This response is the defense-in-depth guard for
        the case where only a mock result exists in such an environment.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Monthly transaction limit reached for the agent's tier.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TransactionAccepted:
      description: |
        The request is valid, policy-approved, and has been queued.
        The transaction will be broadcast by a background worker.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/TransactionAccepted' }
  schemas:
...
    AgentTier:
      type: string
      enum: [FREE, PRO, ENTERPRISE]
...
    Agent:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        apiKeyPrefix: { type: string }
        walletAddress: { type: string }
        ensName:
          type: string
          nullable: true
        chainIds:
          type: array
          items: { type: integer }
        active: { type: boolean }
        tier: { $ref: '#/components/schemas/AgentTier' }
        erc8004:
          type: array
          description: |
            The agent's ERC-8004 identities, one per chain (R2). Minted lazily
            from the agent's own wallet on its first funded ERC-8183 job as a
            provider; empty until then.
          items: { $ref: '#/components/schemas/AgentErc8004Identity' }
        policy:
          type: object
          nullable: true
          allOf:
            - { $ref: '#/components/schemas/Policy' }
        billing:
          type: object
          nullable: true
          properties:
            txCountThisPeriod: { type: integer }
            totalFeesCollectedUsd: { type: string }
            subscriptionActive: { type: boolean }
...
    SimulateSwapRequest:
      type: object
      required: [fromToken, toToken, amountIn]
      properties:
        fromToken: { type: string }
        toToken: { type: string }
        amountIn:
          type: string
          description: Human-readable decimal amount (not wei).
        chainId: { type: integer, default: 1 }
        slippageTolerance:
          type: number
          minimum: 0.01
          maximum: 50
          default: 0.5
    ExecuteSwapRequest:
      allOf:
        - $ref: '#/components/schemas/SimulateSwapRequest'
        - type: object
          required: [simulationId]
          properties:
            simulationId: { type: string }
            idempotencyKey: { type: string }
    TransferRequest:
      type: object
      required: [to, token, amount]
      properties:
        token:
          type: string
          description: '"ETH" or an ERC-20 contract address.'
        to: { type: string }
        amount: { type: string }
        chainId: { type: integer, default: 1 }
        idempotencyKey: { type: string }
    AssetAmountRequest:
      type: object
      required: [asset, amount]
      properties:
        asset: { type: string, description: Token address. }
        amount:
          type: string
          description: Human-readable decimal amount, or "max" for withdraws.
        chainId: { type: integer, default: 1 }
        idempotencyKey: { type: string }
    Erc4626Request:
      type: object
      required: [vault, asset, amount]
      properties:
        vault: { type: string, description: Vault contract address. }
        asset: { type: string, description: Underlying token (for decimals + pricing). }
        amount: { type: string }
        chainId: { type: integer, default: 1 }
        idempotencyKey: { type: string }
...
    SimulationResult:
      type: object
      properties:
        simulationId: { type: string }
        success: { type: boolean }
        gasEstimate: { type: string }
        amountOut: { type: string }
        error:
          type: string
          nullable: true

    # ── Wallet ───────────────────────────────────────────────────────────
...